requirements.txt
asgiref==3.6.0
Django==3.2
django-cors-headers==3.14.0
djangorestframework==3.14.0
psycopg2-binary==2.9.6
python-dotenv==0.21
pytz==2023.3
gunicorn
django-storages
boto3
